Reimporting Project
Reimporting my project to github because of this unsolvable duplication issue. I think it has to do with icloud drive updates. Im on an old mac using a 2019 operating system and I have founds posts online dating back to 2019 of people have the same issues and there being a bug in the OS. Hopefully having the project on my local drive fixes this.